fix(disk): ntriples builds never write into a published generation; post-build save keeps property columns
A disk build on a handle sitting on a published generation wrote the
rebuild into that snapshot in place (stale interner.bin.zst shadowed
the fresh interner.json; reload failed with an unresolved type key),
and a build after any mutation finalised into the mutation workspace
and vanished with it. Bulk builds now stage in a workspace and publish
through the GenerationTxn protocol; published snapshots are pinned
byte-identical before/after in both Rust and Python tests. Also fixed
en route: save() after any ntriples disk build silently dropped every
property column (packed sidecar schema was never grown into the type
schema) - properties now survive the save/reload round trip.
