+++ tests/api-baselines/source-quality.json
@@ -181,13 +181,6 @@
     },
     {
       "path": "crates/kglite/src/graph/io/ntriples/loader.rs",
-      "qualified_name": "crate::finalize_disk_graph",
-      "lines": 308,
-      "branches": 46,
-      "nesting": 5
-    },
-    {
-      "path": "crates/kglite/src/graph/io/ntriples/loader.rs",
       "qualified_name": "crate::flush_entity",
       "lines": 212,
       "branches": 25,
